# backend/src/main/kotlin/cz/loono/backend/db/model/ExaminationRecord.kt
package cz.loono.backend.db.model

import cz.loono.backend.api.dto.ExaminationStatusDto
import cz.loono.backend.api.dto.ExaminationTypeEnumDto
import org.hibernate.Hibernate
import java.time.LocalDateTime
import javax.persistence.Column
import javax.persistence.Entity
import javax.persistence.GeneratedValue
import javax.persistence.GenerationType
import javax.persistence.Id
import javax.persistence.ManyToOne
import javax.persistence.Table

@Entity
@Table(name = "\"examination_record\"")
data class ExaminationRecord(
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    val id: Long = 0,

    @Column(nullable = false, columnDefinition = "TEXT")
    val type: ExaminationTypeEnumDto = ExaminationTypeEnumDto.GENERAL_PRACTITIONER,

    @Column
    val date: LocalDateTime? = null,

    @ManyToOne(optional = false)
    val account: Account = Account(),

    @Column(nullable = false)
    val firstExam: Boolean = true,

    @Column(nullable = false, columnDefinition = "TEXT")
    var status: ExaminationStatusDto = ExaminationStatusDto.NEW
) {
    override fun equals(other: Any?): Boolean {
        if (this === other) return true
        if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
        other as ExaminationRecord

        @Suppress("SENSELESS_COMPARISON")
        return id != null && id == other.id
    }

    override fun hashCode(): Int = 0

    @Override
    override fun toString(): String {
        return this::class.simpleName + "(id = $id , type = $type , date = $date , accountId = ${account.uid}, firstExam = $firstExam, status = $status)" // ktlint-disable max-line-ength
    }
}