# backend/src/main/resources/doc/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Loono API",
        "description": "Loono API for the mobile application. The whole API is JSON based.\n\nFailed API operations return [Error](https://okarmazin.stoplight.io/docs/loono/c2NoOjI2ODI2OA-error-response).\n\nGeneral error codes:\n|HTTP Status|code |description|\n|--- |--- | --- |\n|400|MISSING_PRIMARY_EMAIL| The Firebase login method did not include user email. Email is required and must be saved in Firebase before any API access. |\n|400|MISSING_USER_NAME| The Firebase login method did not include user name. Name is required and must be saved in Firebase before any API access. |",
        "contact": {
            "name": "Loono",
            "url": "https://www.loono.cz/"
        },
        "license": {
            "name": "MIT license",
            "url": "http://www.opensource.org/licenses/mit-license.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://loono.ceskodigital.net:443",
            "description": "Inferred Url"
        }
    ],
...
    "paths": {
...
        "/examinations": {
            "get": {
                "summary": "Get Examinations Status",
                "tags": [
                    "Examinations"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PreventionStatus"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "operationId": "getExaminations",
                "description": "Returns the examination records of the current user.",
                "security": [
                    {
                        "Bearer_Auth": []
                    }
                ]
            },
            "post": {
                "summary": "Create or update examination record",
                "operationId": "post-examinations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExaminationRecord"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer_Auth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExaminationRecord"
                            }
                        }
                    }
                }
            }
        },
        "/examinations/{type}/confirm": {
            "post": {
                "summary": "Confirm Examination",
                "operationId": "completeExamination",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExaminationRecord"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Examinations"
                ],
                "description": "Notify the API that an examination has been performed.\n\nThe date of the completion must not be in the future.",
                "security": [
                    {
                        "Bearer_Auth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExaminationId"
                            }
                        }
                    },
                    "description": ""
                }
            },
            "parameters": [
                {
                    "schema": {
                        "type": "string"
                    },
                    "name": "type",
                    "in": "path",
                    "required": true,
                    "description": "One of the `ExaminationTypeEnum` values"
                }
            ]
        },
        "/examinations/{type}/cancel": {
            "post": {
                "summary": "Cancel Examination",
                "operationId": "cancelExamination",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExaminationRecord"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Examinations"
                ],
                "description": "Notify the API that an examination has been performed.\n\nThe date of the completion must not be in the future.",
                "security": [
                    {
                        "Bearer_Auth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExaminationId"
                            }
                        }
                    },
                    "description": ""
                }
            },
            "parameters": [
                {
                    "schema": {
                        "type": "string"
                    },
                    "name": "type",
                    "in": "path",
                    "required": true,
                    "description": "One of the `ExaminationTypeEnum` values"
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error response",
                "description": "When an API operation fails, API returns this error which may contain a specific error code pertaining to the failed operation.\n\nAll fields are optional and nullable.",
                "type": "object",
                "properties": {
                    "code": {
                        "description": "Machine readable error code. Error codes are defined by the individual API operations. Error code may be used by the client application for UX tailoring.",
                        "type": "string",
                        "nullable": true
                    },
                    "message": {
                        "description": "Human readable error message. These messages are only rarely shown to users and are more often used as a debugging tool for API clients. ",
                        "type": "string",
                        "nullable": true
                    }
                }
            },
            "ExaminationId": {
                "title": "ExaminationId",
                "type": "object",
                "description": "A examination identification.\n\n`month` and `year` are optional and nullable. If any of them is missing or `null`, API defaults to `LocalDate.now()` - the current year and month.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Must not be in the future",
                        "format": "int64",
                        "default": null
                    }
                },
                "required": [
                    "id"
                ]
            },
...
            "ExaminationTypeEnum": {
                "type": "string",
                "title": "ExaminationTypeEnum",
                "description": "",
                "enum": [
                    "BREAST_SELF",
                    "COLONOSCOPY",
                    "DENTIST",
                    "DERMATOLOGIST",
                    "GENERAL_PRACTITIONER",
                    "GYNECOLOGIST",
                    "MAMMOGRAM",
                    "OPHTHALMOLOGIST",
                    "TESTICULAR_SELF",
                    "TOKS",
                    "ULTRASOUND_BREAST",
                    "UROLOGIST",
                    "VENEREAL_DISEASES"
                ]
            },
...
            "ExaminationStatus": {
                "type": "string",
                "title": "ExaminationStatus",
                "description": "Status enum of the examination",
                "enum": [
                    "NEW",
                    "TO_BE_CONFIRMED",
                    "CONFIRMED",
                    "CANCELED",
                    "UNKNOWN"
                ]
            },
            "PreventionStatus": {
                "type": "object",
                "title": "PreventionStatus",
                "description": "Overview status of examinations",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "examinationType": {
                        "$ref": "#/components/schemas/ExaminationTypeEnum"
                    },
                    "intervalYears": {
                        "type": "integer"
                    },
                    "lastExamDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "firstExam": {
                        "type": "boolean"
                    },
                    "priority": {
                        "type": "integer"
                    },
                    "state": {
                        "$ref": "#/components/schemas/ExaminationStatus"
                    },
                    "streak": {
                        "type": "integer"
                    }
                },
                "required": [
                    "examinationType",
                    "intervalYears",
                    "firstExam",
                    "priority",
                    "state",
                    "streak"
                ]
            },
            "ExaminationRecord": {
                "title": "ExaminationRecord",
                "type": "object",
                "description": "Examination object to update or create an object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ExaminationTypeEnum"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/ExaminationStatus"
                    },
                    "firstExam": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "type"
                ]
            }
        },
        "securitySchemes": {
            "Bearer_Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "Basic_Auth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}